adds document metadata to localStorage and UI
Each document now has an optional name, and automatically generated
create and update timestamps. Also, added a delete button. Needed to
lift the open file page state up to the application level to get the
page to update when a document is deleted, but I probably should have
done that before anyway.