adds ability to edit point brightness
Lots of changes to support selection of points and using a slider to
change the brightness. I also managed to get fullscreen working, but
it's still a little flakey, probably when it takes too long to
compute.

wip

fixes preview crash

fullscreen works! (sometimes--needs work)

adds dirty-checking, so it doesn't change updated at on every mouseUp